+++ Braintree.Tests/CreditCardNumbers.cs
@@ -1,0 +1,14 @@
+namespace Braintree.Tests
+{
+    public class CreditCardNumbers
+    {
+        public static readonly string Prepaid         = "4111111111111210";
+        public static readonly string Commercial      = "4111111111131010";
+        public static readonly string Payroll         = "4111111114101010";
+        public static readonly string Healthcare      = "4111111510101010";
+        public static readonly string DurbinRegulated = "4111161010101010";
+        public static readonly string Debit           = "4117101010101010";
+        public static readonly string Unknown         = "4111111111112101";
+        public static readonly string No              = "4111111111310101";
+    }
+}